pyproject.toml
[project]
name = "media-icopy"
version = "1.0.0"
description = "Copy media files from iPhone to Windows PC via MTP"
requires-python = ">=3.10"
classifiers = [
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pywin32==311",
    "pyyaml==6.0.3",
    "customtkinter==5.2.2",
    "pillow==12.2.0",
    "pillow-heif==1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pyinstaller>=6.0",
]

[project.scripts]
media-icopy-gui = "src.gui.main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

